support for py2neo 2.0.8 + fix for test auth

+++ tests/conftest.py
@@ -6,6 +6,7 @@ warnings.simplefilter("error")
 
 @pytest.fixture(scope='session')
 def metadata():
+    from py2neo.error import GraphError
     from neolixir import metadata
     if 'NEO4J_TEST_SERVER' in os.environ:
         metadata.url = 'http://{0}/db/data/'.format(os.environ['NEO4J_TEST_SERVER'])
@@ -17,7 +18,15 @@ def metadata():
             metadata.authenticate('neo4j', 'neo4j')
     except ImportError:
         pass
-    metadata.graph.delete_all()
+    try:
+        metadata.graph.delete_all()
+    except GraphError as e:
+        if str(e).find('response 403') > 0:
+            metadata.change_password('neo4j', 'neo4j', 'testpass')
+            metadata.change_password('neo4j', 'testpass', 'neo4j')
+            metadata.graph.delete_all()
+        else:
+            raise e
     import models
     metadata.init()
     return metadata
